Fixed latest version details API
In APIs which are used to send new version updates availability and new version details for products was having issue for comparing version becuase it was comparing version as normal strings. This commit makes the changes in the code to use PHP's version_compare method to compare version correctly and return correct responses.